+++ lib/xcodeproj/constants.rb
@@ -453,5 +453,12 @@ module Xcodeproj
       $(inherited)
       ${inherited}
     ).freeze
+
+    # @return [Hash] Possible types for a scheme's 'ExecutionAction' node
+    #
+    EXECUTION_ACTION_TYPE = {
+      :shell_script  => 'Xcode.IDEStandardExecutionActionsCore.ExecutionActionType.ShellScriptAction',
+      :send_email    => 'Xcode.IDEStandardExecutionActionsCore.ExecutionActionType.SendEmailAction',
+    }.freeze
   end
 end