+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ##### Enhancements
 
-* None.  
+* Add support for pre/post-actions in scheme actions  
+  [thiagohmcruz](https://github.com/thiagohmcruz)
+  [#401](https://github.com/CocoaPods/CocoaPods/issues/401)
 
 ##### Bug Fixes